A fix to get rid of a preconcurrency requirement
I no longer need to wait for a MainActor update before responding to a
put fueling message.  Also fixed were some run time error where a non
sendable item was not on the MainActor.

However, The watch is no longer receiving application context updates.
I see where one was requested and on the phone I see that a context
update was sent.