Force app context updates to be sent
Add a timestamp to app context messages to guarantee each update is
unique and therefore sent to the watch.  Otherwise the app will not
resend previously sent updates.  That is bad if the watch wasn't ready
to receive the update when it was sent.

The watch wil ignore the timestamp.