Do not import SwiftData when not needed
Not only a good idea but a source of compile errors in Xcode16b2 in this
code. Not sure why as a simple test app compiles views fine when
SwiftData is imported.  Possibly a side effect of the way I am using
SwiftData in this app, i.e. not tied to SwiftUI