Work on Fuel Entry Updates
The code was working, but database updates were happening outside
of the reducer. This code fixes that. It also moves what was state
methods into the reducer so all state updates are not only triggered by
the reducer, but are updated by the reducer, too.

Except for database updates... they still live in FuelingDB.