More watch app testing
Changed FuelEntryView to use plain buttons instead of Text items with a
tap gesture. That was done in an attempt to get around a test framework
issue (can't tap) but didn't help.

The can't tap issue has been reported many times over the years. Not
sure which of the suggested workarounds I'll try.